Move cars to ECS (first steps)

+++ Ported/sschoener/MagneticRoads/Assets/Scripts/Intersection.cs
@@ -9,8 +9,8 @@ public static class Intersections
 	public static int3[] Index;
 	public static float3[] Normal;
 
-	public static List<int>[] Neighbors;
-	public static List<int>[] NeighborSplines;
+	public static List<ushort>[] Neighbors;
+	public static List<ushort>[] NeighborSplines;
 	
 	// each side (top/bottom) has its own "occupied" flag
 	// (a car on the underside doesn't block a car on the top)
@@ -22,13 +22,13 @@ public static class Intersections
 		Position = new float3[n];
 		Index = new int3[n];
 		Normal = new float3[n];
-		Neighbors = new List<int>[n];
-		NeighborSplines = new List<int>[n];
+		Neighbors = new List<ushort>[n];
+		NeighborSplines = new List<ushort>[n];
 		Occupied = new OccupiedSides[n];
 		for (int i = 0; i < n; i++)
 		{
-			Neighbors[i] = new List<int>();
-			NeighborSplines[i] = new List<int>();
+			Neighbors[i] = new List<ushort>();
+			NeighborSplines[i] = new List<ushort>();
 		}
 	}
